Allow android multi touch actions without an element
