+++ lib/devices/android/android-controller.js
@@ -1004,18 +1004,26 @@ androidController.tap = function (elementId, x, y, count, cb) {
     this.proxy(opts, loop);
   }.bind(this);
 
-  if (x !== 0 || y !== 0) {
-    this.proxy(["element:getLocation", {elementId: elementId}], function (err, res) {
-      if (err) return cb(err);
-      var value = res.value;
-      x += parseInt(value.x);
-      y += parseInt(value.y);
+  if (elementId) {
+    // we are either tapping on the default location of the element
+    // or an offset from the top left corner
+    if (x !== 0 || y !== 0) {
+      this.proxy(["element:getLocation", {elementId: elementId}], function (err, res) {
+        if (err) return cb(err);
+        var value = res.value;
+        x += parseInt(value.x);
+        y += parseInt(value.y);
 
-      opts = ["click", {x: x, y: y}];
+        opts = ["click", {x: x, y: y}];
+        loop();
+      }.bind(this));
+    } else {
+      opts = ["element:click", {elementId: elementId}];
       loop();
-    }.bind(this));
+    }
   } else {
-    opts = ["element:click", {elementId: elementId}];
+    // we have absolute coordinates
+    opts = ["click", {x: x, y: y}];
     loop();
   }
 };
@@ -1156,6 +1164,7 @@ androidController.parseTouch = function (gestures, cb) {
           x: (gesture.options.x || 0),
           y: (gesture.options.y || 0)
         };
+
         touchStateObject = {
           timeOffset: 0.005,
           touch: tapPoint
@@ -1221,11 +1230,19 @@ androidController.performMultiAction = function (elementId, actions, cb) {
   }.bind(this), function (err) {
     if (err) return cb(err);
 
-    var opts = {
-      elementId: elementId,
-      actions: states
-    };
-    return this.proxy(["element:performMultiPointerGesture", opts], cb);
+    var opts;
+    if (elementId) {
+      opts = {
+        elementId: elementId,
+        actions: states
+      };
+      return this.proxy(["element:performMultiPointerGesture", opts], cb);
+    } else {
+      opts = {
+        actions: states
+      };
+      return this.proxy(["performMultiPointerGesture", opts], cb);
+    }
   }.bind(this));
 };
